+++ rich/filesize.py
@@ -13,10 +13,17 @@ See Also:
 
 __all__ = ["decimal"]
 
-from typing import Iterable, List, Tuple
+from typing import Iterable, List, Tuple, Optional
 
 
-def _to_str(size: int, suffixes: Iterable[str], base: int) -> str:
+def _to_str(
+    size: int,
+    suffixes: Iterable[str],
+    base: int,
+    *,
+    precision: Optional[int] = 1,
+    separator: Optional[str] = " ",
+) -> str:
     if size == 1:
         return "1 byte"
     elif size < base:
@@ -26,7 +33,12 @@ def _to_str(size: int, suffixes: Iterable[str], base: int) -> str:
         unit = base ** i
         if size < unit:
             break
-    return "{:,.1f} {}".format((base * size / unit), suffix)
+    return "{:,.{precision}f}{separator}{}".format(
+        (base * size / unit),
+        suffix,
+        precision=precision,
+        separator=separator,
+    )
 
 
 def pick_unit_and_suffix(size: int, suffixes: List[str], base: int) -> Tuple[int, str]:
@@ -38,7 +50,12 @@ def pick_unit_and_suffix(size: int, suffixes: List[str], base: int) -> Tuple[int
     return unit, suffix
 
 
-def decimal(size: int) -> str:
+def decimal(
+    size: int,
+    *,
+    precision: Optional[int] = 1,
+    separator: Optional[str] = " ",
+) -> str:
     """Convert a filesize in to a string (powers of 1000, SI prefixes).
 
     In this convention, ``1000 B = 1 kB``.
@@ -50,6 +67,8 @@ def decimal(size: int) -> str:
 
     Arguments:
         int (size): A file size.
+        int (precision): The number of decimal places to include (default = 1).
+        str (separator): The string to separate the value from the units (default = " ").
 
     Returns:
         `str`: A string containing a abbreviated file size and units.
@@ -57,6 +76,14 @@ def decimal(size: int) -> str:
     Example:
         >>> filesize.decimal(30000)
         '30.0 kB'
+        >>> filesize.decimal(30000, precision=2, separator="")
+        '30.00kB'
 
     """
-    return _to_str(size, ("kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"), 1000)
+    return _to_str(
+        size,
+        ("kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"),
+        1000,
+        precision=precision,
+        separator=separator,
+    )