+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.6.0] - Unreleased
+
+### Added
+
+- Added precision argument to filesize.decimal
+- Added separator argument to filesize.decimal
+
 ## [10.5.0] - 2021-05-07
 
 ### Fixed